Replace hidden mobile filter sheet with always-visible inline filter bar
- Removes the mobile bottom sheet and desktop sidebar filter (BillsFilter)
- New FilterQuickAccess component: all 10 filters always visible inline
  - Title/bill# text inputs + sponsor typeahead
  - 7 wrapping dropdown pills: Status, Congress, Policy area, State, Introduced, Last action, Type
  - Active chips with per-filter remove + Clear all
- Extracted SponsorCombobox into its own file
- Moved shared filter constants to lib/constants/filters.ts
- Removed pending-filter staging (desktop Apply workflow no longer needed)
- Deleted bills-filter.tsx (no remaining imports)
